refactor : sidebar 수정 및 Config 수정
