+++ src/main/resources/templates/sidebar.html
@@ -7,7 +7,7 @@
             <a th:href="@{/}">
                 <i class="fa fa-home fa-2x"></i>
                 <span class="nav-text">
-                                home
+                                Home
                             </span>
             </a>
         </li>
@@ -52,7 +52,7 @@
             <a href="https://github.com/woominsik">
                 <i class="fa fa-github fa-2x"></i>
                 <span class="nav-text">
-                                github
+                                Github
                     </span>
             </a>
         </li>
@@ -60,6 +60,15 @@
 
 
     <ul class="logout">
+        <li>
+            <a href="#">
+                <i class="fa fa-user-circle fa-2x"></i>
+                <span class="nav-text">
+                                SignUp
+                </span>
+            </a>
+        </li>
+
         <li>
             <a href="#">
                 <i class="fa fa-power-off fa-2x"></i>
